requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-ckeditor==5.3.0
django-js-asset==0.1.1
docopt==0.4.0
idna==2.6
mailchimp==2.0.9
olefile==0.44
Pillow==4.3.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
